LayoutTests/media/media-source/media-source-append-b-frame-tail-overshoot-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer('video/mp4; codecs="avc1.4d401e"'))
RUN(sourceBuffer.appendBuffer(MP4.concat(init, segment1)))
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0.1') OK
EXPECTED (sourceBuffer.buffered.end(0) == '0.2') OK
RUN(sourceBuffer.appendBuffer(segment2))
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0') OK
EXPECTED (sourceBuffer.buffered.end(0) == '0.2') OK
END OF TEST
